APEXMALHAR-2093 Removed usages of Idempotent Storage Manager
