+++ contrib/src/main/java/com/datatorrent/contrib/redis/AbstractRedisInputOperator.java
@@ -25,6 +25,9 @@ import java.util.List;
 
 import javax.validation.constraints.NotNull;
 
+import org.apache.apex.malhar.lib.wal.FSWindowDataManager;
+import org.apache.apex.malhar.lib.wal.WindowDataManager;
+
 import redis.clients.jedis.ScanParams;
 import redis.clients.jedis.ScanResult;
 
@@ -32,7 +35,6 @@ import com.datatorrent.api.Operator.CheckpointListener;
 import com.datatorrent.api.Context.OperatorContext;
 import com.datatorrent.netlet.util.DTThrowable;
 import com.datatorrent.lib.db.AbstractKeyValueStoreInputOperator;
-import com.datatorrent.lib.io.IdempotentStorageManager;
 
 /**
  * This is the base implementation of a Redis input operator.
@@ -57,7 +59,7 @@ public abstract class AbstractRedisInputOperator<T> extends AbstractKeyValueStor
   private transient boolean skipOffsetRecovery = true;
 
   @NotNull
-  private IdempotentStorageManager idempotentStorageManager;
+  private WindowDataManager windowDataManager;
 
   private transient OperatorContext context;
   private transient long currentWindowId;
@@ -83,7 +85,7 @@ public abstract class AbstractRedisInputOperator<T> extends AbstractKeyValueStor
     recoveryState = new RecoveryState();
     recoveryState.scanOffsetAtBeginWindow = 0;
     recoveryState.numberOfScanCallsInWindow = 0;
-    setIdempotentStorageManager(new IdempotentStorageManager.NoopIdempotentStorageManager());
+    setWindowDataManager(new FSWindowDataManager());
   }
 
   @Override
@@ -92,7 +94,7 @@ public abstract class AbstractRedisInputOperator<T> extends AbstractKeyValueStor
     currentWindowId = windowId;
     scanCallsInCurrentWindow = 0;
     replay = false;
-    if (currentWindowId <= getIdempotentStorageManager().getLargestRecoveryWindow()) {
+    if (currentWindowId <= getWindowDataManager().getLargestRecoveryWindow()) {
       replay(windowId);
     }
   }
@@ -105,11 +107,11 @@ public abstract class AbstractRedisInputOperator<T> extends AbstractKeyValueStor
       if (!skipOffsetRecovery) {
         // Begin offset for this window is recovery offset stored for the last
         // window
-        RecoveryState recoveryStateForLastWindow = (RecoveryState) getIdempotentStorageManager().load(context.getId(), windowId - 1);
+        RecoveryState recoveryStateForLastWindow = (RecoveryState) getWindowDataManager().load(context.getId(), windowId - 1);
         recoveryState.scanOffsetAtBeginWindow = recoveryStateForLastWindow.scanOffsetAtBeginWindow;
       }
       skipOffsetRecovery = false;
-      RecoveryState recoveryStateForCurrentWindow = (RecoveryState) getIdempotentStorageManager().load(context.getId(), windowId);
+      RecoveryState recoveryStateForCurrentWindow = (RecoveryState) getWindowDataManager().load(context.getId(), windowId);
       recoveryState.numberOfScanCallsInWindow = recoveryStateForCurrentWindow.numberOfScanCallsInWindow;
       if (recoveryState.scanOffsetAtBeginWindow != null) {
         scanOffset = recoveryState.scanOffsetAtBeginWindow;
@@ -153,7 +155,7 @@ public abstract class AbstractRedisInputOperator<T> extends AbstractKeyValueStor
   {
     super.setup(context);
     sleepTimeMillis = context.getValue(context.SPIN_MILLIS);
-    getIdempotentStorageManager().setup(context);
+    getWindowDataManager().setup(context);
     this.context = context;
     scanOffset = 0;
     scanComplete = false;
@@ -161,7 +163,7 @@ public abstract class AbstractRedisInputOperator<T> extends AbstractKeyValueStor
     scanParameters.count(scanCount);
     
     // For the 1st window after checkpoint, windowID - 1 would not have recovery
-    // offset stored in idempotentStorageManager
+    // offset stored in windowDataManager
     // But recoveryOffset is non-transient, so will be recovered with
     // checkPointing
     // Offset recovery from idempotency storage can be skipped in this case
@@ -181,9 +183,9 @@ public abstract class AbstractRedisInputOperator<T> extends AbstractKeyValueStor
     recoveryState.scanOffsetAtBeginWindow = scanOffset;
     recoveryState.numberOfScanCallsInWindow = scanCallsInCurrentWindow;
 
-    if (currentWindowId > getIdempotentStorageManager().getLargestRecoveryWindow()) {
+    if (currentWindowId > getWindowDataManager().getLargestRecoveryWindow()) {
       try {
-        getIdempotentStorageManager().save(recoveryState, context.getId(), currentWindowId);
+        getWindowDataManager().save(recoveryState, context.getId(), currentWindowId);
       } catch (IOException e) {
         DTThrowable.rethrow(e);
       }
@@ -194,7 +196,7 @@ public abstract class AbstractRedisInputOperator<T> extends AbstractKeyValueStor
   public void teardown()
   {
     super.teardown();
-    getIdempotentStorageManager().teardown();
+    getWindowDataManager().teardown();
   }
 
   /*
@@ -231,7 +233,7 @@ public abstract class AbstractRedisInputOperator<T> extends AbstractKeyValueStor
   public void committed(long windowId)
   {
     try {
-      getIdempotentStorageManager().deleteUpTo(context.getId(), windowId);
+      getWindowDataManager().deleteUpTo(context.getId(), windowId);
     } catch (IOException e) {
       throw new RuntimeException("committing", e);
     }
@@ -240,16 +242,16 @@ public abstract class AbstractRedisInputOperator<T> extends AbstractKeyValueStor
   /*
    * get Idempotent Storage manager instance
    */
-  public IdempotentStorageManager getIdempotentStorageManager()
+  public WindowDataManager getWindowDataManager()
   {
-    return idempotentStorageManager;
+    return windowDataManager;
   }
 
   /*
    * set Idempotent storage manager instance
    */
-  public void setIdempotentStorageManager(IdempotentStorageManager idempotentStorageManager)
+  public void setWindowDataManager(WindowDataManager windowDataManager)
   {
-    this.idempotentStorageManager = idempotentStorageManager;
+    this.windowDataManager = windowDataManager;
   }
 }
