+++ library/src/main/java/org/apache/apex/malhar/lib/state/managed/IncrementalCheckpointManager.java
@@ -32,7 +32,7 @@ import javax.validation.constraints.NotNull;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import org.apache.apex.malhar.lib.wal.WindowDataManager;
+import org.apache.apex.malhar.lib.wal.FSWindowDataManager;
 
 import com.google.common.base.Preconditions;
 import com.google.common.base.Throwables;
@@ -49,7 +49,7 @@ import com.datatorrent.netlet.util.Slice;
  *
  * This component is also responsible for purging old time buckets.
  */
-public class IncrementalCheckpointManager extends WindowDataManager.FSWindowDataManager
+public class IncrementalCheckpointManager extends FSWindowDataManager
     implements ManagedStateComponent
 {
   private static final String WAL_RELATIVE_PATH = "managed_state";
